Updated Multi Singularity
This is the gotta go fast version of Multi Singularity, which is named like that because the special numbers have a 10x boost in regards to multiplying power. The cost scaling is also quartered, meaning you can get very far very quickly. I think you might even not need to buy more computer power to complete it.